backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
requests
beautifulsoup4
pillow
pytesseract
PyMuPDF
sentence-transformers
faiss-cpu
matplotlib
google-generativeai
google-cloud-aiplatform
PyMuPDF